Always redirect to My Files after upload
Summary:
I've found it quite confusing that uploading a single image displays the image itself but uploading more images displays My Files.
I've also got a user report about it because most users don't know that they can drop the image directly to the comment textarea and they are interested mainly in the ID of the uploaded file.

Test Plan: Upload a file.

Reviewers: epriestley

Reviewed By: epriestley

CC: aran, epriestley

Differential Revision: https://secure.phabricator.com/D1891